Add useDomainFromUrlParam util
